selectIsHealer refactored to selectors
